Change CTRLZ to CTRLX to background the process
Previously you had to type ctrl-z two times to put the process into
the background.  I am thinking the first was the expect script and
the second was the neo perl script, but either way that's how it
worked.

The problem with that mechanism was that sometimes in my remote
shell I was in vim and needed to put vim into the background of my
remote shell only (without putting the whole session into the back-
ground of my local shell).  Basically I had no way to put the remo-
te vim process into the background of the remote shell.

So to get around this I have changed the neo script to use CTRLX
instead of CTRLZ.  This also means having to change \032 to \030.
The new way to put the remote shell into the background is to type
ctrl-x and then ctrl-z.