Fix screen modifier/remote bash execute bug
The remote bash execute waits for a hash symbol '#' before running
the exec bash line and subsequently the line that sets the PS1 en-
vironment variable.  However, when running with the screen modifier
there were some systems which would never show a '#' symbol because
the remote system root account doesn't have a .screenrc file and
thus displays a welcome message instead of simply dropping the user
to a prompt.

What would happen is, it would wait forever for the '#' and then
you would have to cancel out of it, ssh to the remote computer and
then run ``screen -r'', or else the screen would just be waiting
there forever.

This fix modifies the regular expression to match before running
the exec bash line.  In addition to succeeding on the '#' symbol,
it will also succeed when it sees the text ``Press Space or Return
to end''.