Save pix-passwords file to local cache
This should allow faster pix logins since the passwords are saved
locally and the script doesn't have to ssh to pinky everytime to
grab them.

If you haven't noticed by now, this script needs a symlink
a -> /nothing-to-see-here-move-along in your home folder on pinky.

Since the format of the pix-passwords file is different than the
formats of the other passwords files, the pix search gets its own
way of printing.  When the print is done it exits.