Convert some (but not all) parsing to C
 * create neo.c
 * rename neo to neo.pl
 * create Makefile
  -you can make install/clean/uninstall
 * create neo-help.pl
  -this is where the help information is now located
 * move the exec of apg into usage.c from neo.pl

Start using C for some stuff if possible.  I'm not really sure if
I know how to continue, but this is at least a start and it works
except for the --showpasswords and the --screen sub-options from
the original neo perl script.  However, they're still available by
running neo.pl --showpasswords search etc.  If they weren't then
logins would not be possible becasue this is the only way to show
passwords now.

The main thing being done in C is parsing command line arguments
and subsequently running some other script in order to accomplish
the function requested by the user.  So for example, if the user
runs ``neo help'' then what actually gets executed is a script
which would print out the help text.