Add hostnames and IP addresses to bash completion
 * Hostnames: neo search . | grep -v ^- | awk '{print $2}'
 * IP addresses: neo search . | grep -v ^- | awk '{print $3}'

This allows the user to type ``neo search 10.2.75.[tab]'' and see
a listing of any match in the second or third field of any of the
password files.  A future issue is probably going to arise because
of the fact that not all password files have the same number of
fields, but I'll deal with that later.

Also it is probably important to know not all the hostnames in the
pinky files are even valid in DNS anyway, so this is really not
going to help in all situations.  But it's cool and useful for lots
of cases so it's being added.