Add --help and --version into the opts variable
Previous tab completion was just for commands only, i.e. only whole
words and did not support any type of command beginning with ``--''.
Since one would expect to be able to type ``neo --help'' or
``neo --version'' this is included in neo.c to be able to handle
these types of arguments on the command line.  However, I forgot to
add them to the completion so here they are.