Make the search option variables globally exported
I want to be able to have tab completion for every host in any of
the pinky files and for any of the IP addresses in any of the pinky
files.  So to do that one way is to grep and pipe to awk and do lots
of other really intensive things on the file data until you get the
hostnames or the IP addresses.

Previously though, this was getting run every time you press the tab
key, so it was just going really slowly to get the results up.  Now
though, I'm putting them outside the function call so they just get
run one time when the file gets sourced originally.  The variables
are exported (yes it makes a mess), but to me it's worth it in order
to have the extra speed every time I press tab.