Split off search functions into new file search.pl
 * Move the search and save_local_cache subs into search.pl
 * In search.pl, act based on numeric value of first argument
  - 128 is a search without showing passwords
  - 192 is a search showing passwords
 * Remove Getopt::Lucid calls for search and show_passwords
 * In search.c, call search.pl instead of neo.pl
  - Use numbers to specify the action to take

The purpose is to make things more separated rather than having
everything inside the neo.pl file being called directly from one
of the .c files.  Now it's hoped that having a few different scripts
would make it easier for someone to come along and figure things
out without much effort.

To do: fix up the Makefile to install the search.pl into
/usr/local/bin along with the others.

An added thing to cleaning up is to not have to install the
Getopt::Lucid perl module since most distributions don't have it
by default.  So now using 192 or 168 as argv[1]---or in perl it's
argv[0] because perl doesn't pass in the program name---there
exists a search.pl which is entirely module free.

Originally I tried to do #define statements so I could pass in the
options with a bitwise AND.  However, this wasn't too well thought
out, and when I tried to make it work I pretty much got confused
and gave up.  If I end up using defined things, or const variables
or whatever, it would pretty much get rid of the need for the
search_opts structure.  So I guess I have to choose one or the
other.