Fix a small issue
When searching everything was fine, but when logging in, the neo.pl
script was calling get_info function and was trying to grab the
output of `neo.pl search --showpasswords query` which didn't work
because neo.pl doesn't have a search sub any more.  So now instead
it calls the correct script which is search.pl and gives the
correct argument which is 192 since we're needing to get passwords.