Introduce a check_ip scanner to validate IPs
Using the re2c package from Slackbuilds.org, it's possible to
write a file which re2c then converts into C code and which I can
use in my program.  The advantage is I guess that C is faster than
expect/tcl/perl?  No it was just for fun I guess.  The re2c
makers call the result a scanner but I called my function that it
made check_ip().  It returns a zero on success and one on failure.

The way it is done is to write a .re file (a regular expression
file).  This consists of a lot of stuff I didn't really take the
time to learn plus a section where you define your different
regexes along with an action to take for when you get a match.
It looks very reminiscent/similar to how expect does some things.
In my case, I just had it make a function check_ip(char *) which
I feed the IP into as a string.  This calls the scan function
which returns a 0 if a match is made on the string or otherwise
returns a 1.  My script is really just a wrapper, and I'm not
sure how I'm supposed to handle it if I wanted to do another
regex without having a conflicting definition for the scan()
function.

neo ssh, pix, and rdp now each check before calling their expect
or perl scripts.  If the IP doesn't match the regex the program
just says ip not valid or something simple.

The regular expression syntax is not 100% PCRE or really is not
even close.  I had to make a lot of changes to my previous regexes.
\d is really [0-0] and \. is now just a '.' character.  I was
still able to (I think) achieve the same regex it just looks a lot
different when written with the new syntax which is in the man page.