Get rid of useless computername variable
In neo-rdp.pl there is a remote desktop caller subroutine which
is passed two variables: the ip and the password.  It then calls
remote desktop to the ip, with user sqltest and the password supp-
lied.

Previously, a computername variable was passed so the rdesktop
could be given a user like web1\sqltest and could therefor assure
that the login was a local login as opposed to a domain-credential
login.  This was because with most customers, the default login
for a remote desktop assumes that you're wanting to supply a domain
level user and password.  So when you type web1\sqltest, it forces
the remote machine to treat it as the local sqltest as opposed to
the domain controller sqltest.  This was being searched for and
retrieved by the get_info subroutine.

So now I'm just hard-coding ip ranges that have a doamin account
I know how to search for.  Then I have a whole other search for
finding those credentials which is also hard-coded.  I'll have to
change this later on since the domain credentials are inline with
each entry in the pinky files but for now it's okay.  Anyway, I
don't need the computername any more which is why it's being disc-
arded.