First attempt at a config file (and parser)
 * neoconfig - the config file
 * config.c  - the functions for parsing neoconfig
 * config.h  - function declarations and a struct declaration
 * Makefile  - add config.o to key places

In config.h, create a generic struct for holding configuration
options from the config file.  Currently the config.c declares a
bunch of structs, one for ssh, one for rdp, etc.  I'm not sure if
this is a good idea yet.  The struct has a bunch of binary values
to indicate whether or not to use uber, dco, whether or not to
establish a tty when doing an ssh command, etc.  Then after those
binary values are character strings which would hold the uber/dco
user name and password and any default args or any extra args
the user wants to pass when running on the command line.

I haven't done this yet, but am hoping to pass the extra args by
appending two minus signs to the end of the command line which is
I think the way they do the same thing when running startx.  E.g.
I think you can say something like startx :1 -- xterm -opt1 -opt2.
So in this case if I wanted RDP v.4 instead of v.5 (used for
remoting into a Win2k8 server) I would do something like:
``neo rdp 4.2.2.2 -- -4''.  In that case, the ``-4'' gets passed
directly to the /usr/bin/rdesktop command line.  It would also be
pretty cool to be able to specify an argument like ``-u'' and have
it override whatever is the default or whatever is in the config
file.

The config file will have the same format as the one for git with
the exception that in this case, some config file lines (such as
the ones for specifying the default argument list to ssh) will be
very long and will need to be split across multiple lines by using
a backslash character.  So there is an escape sequence introduced
that if you type two backslash characters together, then you are
specifying a backslash in the argument string.

[ssh]
    args = -kt24aqx \
    -o StrictHostKeyChecking=no \
    -o Compression=no \
    -o CheckHostIP=no \
    -o PasswordAuthentication=yes \
    -o PubkeyAuthentication=no \
    -o RhostsRSAAuthentication=no \
    -o RSAAuthentication=no

[rdp]
    args = -a 16 -g 1024x768 2>/dev/null

[vnc]
    args = -geometry 1024x768 2>/dev/null

There is a function opt_xxx for each of the headings in square
brackets.  The parser calls the heading opt.  Then it goes and
finds the val and the var.  Once it has all three opt, val, and
var it executes a handle_option function which has some smarts
to figure out which of the opt_xxx functions should be executed.
This opt_xxx function is what is going to take care of setting
up all the stuff inside the opts struct.  The only reason I did
it this way is because I was looking at the cli parser and just
copy/pasted the way it handles searching through the argv list.
Of course, almost all of this is ripped off from git anyway.