Login to pix device with uber credentials
Before the password for each firewall was the same.  Now they have
a new more secure way which is to use the password that is stored
in the uber system.  So now instead of static passwords, they are
the password for whichever user you login with.  Also, the enable
password is the uber password as well.