Include expect the proper way
Since there is not a libexpect symlink in /usr/lib64 on my
computer, I need to include the path with -I (I think) and I need
to include the name of the library with the -L.  I tried with -I
and a lower case -l, but that didn't work.  When I tried with the
upper case -L it worked.  I never did try without any -I, maybe
some other time.