Allow ssh to server as regex, not just IP
The bash-completion now is doing things as a regular expression
which would be matched against the password file.  So for example,
neospire.*ns1.*ip_address gets completed.  The regex could be
matched against a unique line in the password file (hopefully).

This isn't setup yet in the builtin-ssh.c file.  It has to be
called either with 128 or 192 as the first argument on the
command line.  This option needs to be added to builtin-ssh.c in
order for it to be called correctly.