Save argv[1] in case this is a regex login
I am not sure if this is the most efficient or straightforward
way of doing this.  What I want is to save ip and qry just in
case the command turns out to need either of these later on.

If I only set the variable when the need arises then I get a
compiler warning about how I'm not setting the variable.  Since
that is the case I want to go ahead and set the variable anyway,
and if it's not needed then just don't use it.

In this case, the qry variable would either be a regular expression
tabbed out via tab completion, or it would be the string "ip" or
the string "--ip".  If it's not either of the latter, then the
cmd_ssh function treats it as a regex and sends it off to the
neo-ssh.exp program as the last argument.

If it does turn out to that the user has typed --ip or ip, then
this isn't used any further.