Setup to run based on subcommand
This has a few changes.  First it loops through all cli arguments
starting at argv[1].  If it matches one of the arguments to either
ip or --ip, then it looks for the next argument and sets that to
the ip address.  It also makes a note of this and sends 192 to the
neo-ssh.exp expect script so it knows we're doing an ip based login,
and not a regex based login.

It also updates the way the ip itself is checked against the ip
regex matcher, to validate that it's a properly written ip address.
I.e. 555.555.555.555 should generate an error and the program
should exit.  If the opt.ip variable is not set to one/true, then
it doesn't run the ip against the regex.