Simplify the search function
Previously, the function would loop through all the cli arguments
and set options based on what it would find.  So if it found an
argument showpasswords, then it would set the showpasswords bit in
the opt variable.  This seems a bit complicated.  It's easier just
to perform the exec inside the loop and not have that extra
variable in the first place.  So now there is no struct, it just
looks to see if you want to show the passwords and if so runs the
expect script with the 192 argument instead of 128.