Find the ip in a more standard way
When looking for the ip, the next argument after the ip argument
would be the ip itself.  Before, I was just saying argv[2] since I
know argv[2] is really the only place the ip address could be at
anyway.  However, this is more correct since it counts to the next
relative argument instead of blindly choosing argv[2] each time.