Setup a char pointer for the remote command
This is the command to pass to neo-ssh.exp, which will get run on
the remote computer.  On the command line, this is anything after
the double minus sign syntax.  So neo ssh 4.2.2.2 -- ls -alh will
login to the remote computer using the password from pinky and run
ls -alh without allocating a pseudo-tty.