Put cmd_rdp in its own file builtin-rdp.c
 * Cycle through all cli args
 * All args following -- are passed directly to rdesktop itself

In this function, I'm trying to differentiate between when the
user puts an IP address versus when the user is just doing a kind
of hybrid regex type of thing.  If they're doing an IP, I force
them to put the --ip flag followed by the ip.  Otherwise its a
regex.

If they don't put the --ip flag, but they *do* put an IP (this is
verified with my regular expression in check-ip.re), then the
program exits with an error saying that if they want to specify
an ip address they will need to go ahead and put the --ip flag.

If they put the --ip flag, the same regex check runs and makes
absolutely sure it matches the ip address format.  If not, the
program exits with an error message.

Everything after the -- argument is put into its own variable.
This is passed along to the neo-rdp.exp script as trailing stuff
after the ip/regex argument.  That program has some smarts built
into it which basically look at all the trailing arguments (if
there are any) and passes them directly to rdesktop after the
-u sqltest set of arguments.

If we're doing an ip search, we use 128 as the command.  If we're
doing a regex search, we use 64 as the command.