Translate the neo-rdp script from perl to tcl
This took a bit of tinkering but I think I finally have something
that works very well.  I guess one of the main reasons for transl-
ating this was because the neo-ssh.exp already has working code
for grabbing arguments and converting them into something that can
be passed directly to a spawn/exec type of statement.