Clean up the neo-completion script
This script creates the nice tab completion available for some of
the neo commands.  In order to have tab completion for the differ-
ent servers here, I would need to put those servers all into some
variable accessible to the shell at the time the person is typing
in the command.  Since there are so many servers this was creating
a huge memory footprint in my shell's memory area.  I realized
later on that using a bunch of memory for shell variables like I
was doing would later cause problems for commands with really long
argument lists.  Basically it looks like when you run a command
all the arguments are limited to this same memory space, so if you
have already used all the memory with your environment variables,
as I had, then you won't be able to run simple commands like pipes
and xargs type commands etc.  So in order to free up some memory
I'm just going to greatly simplify this whole thing and clean it
up a little bit.