Remove [uber] section from neoconfig
I recently found out from james smith that my uber password is the
same thing as my DCO password.  So for all this time I have had
two sections in the config file with really no need.  Later I will
have to recode the builtin-config.c or whatever, but for now I'm
just going to get rid of the options in the config file.