Disallow ssh/rdp admin logins due to Password Manager Pro
From now on we're going to use a new password tool which does not
have any kind of cli access capability that I know how to use.
All password lookups are going to be through the web interface.
So because of that, the easiest way to break myself of the habit
of using neo to login to remote computers is to just disable the
two functions that allow me to do so.  I can still search though,
but at some point I need to change all the passwords to asterisks
or something like that in order to erase them without erasing the
file format.